refactor: add ProposalActionRolesModule
