+++ src/proposals/proposal-actions/proposal-action-roles/models/proposal-action-role-member.model.ts
@@ -7,7 +7,7 @@ import {
   PrimaryGeneratedColumn,
   UpdateDateColumn,
 } from "typeorm";
-import { User } from "../../../users/models/user.model";
+import { User } from "../../../../users/models/user.model";
 import { ProposalActionRole } from "./proposal-action-role.model";
 
 @ObjectType()
