+++ src/proposals/proposal-actions/proposal-action-roles/proposal-action-roles.resolver.ts
@@ -1,0 +1,4 @@
+import { Resolver } from "@nestjs/graphql";
+
+@Resolver()
+export class ProposalActionRolesResolver {}
