+++ src/proposals/proposals.module.ts
@@ -5,6 +5,7 @@ import { ImagesModule } from "../images/images.module";
 import { RolesModule } from "../roles/roles.module";
 import { VotesModule } from "../votes/votes.module";
 import { Proposal } from "./models/proposal.model";
+import { ProposalActionRolesModule } from "./proposal-actions/proposal-action-roles/proposal-action-roles.module";
 import { ProposalActionsModule } from "./proposal-actions/proposal-actions.module";
 import { ProposalsResolver } from "./proposals.resolver";
 import { ProposalsService } from "./proposals.service";
@@ -14,6 +15,7 @@ import { ProposalsService } from "./proposals.service";
     TypeOrmModule.forFeature([Proposal]),
     forwardRef(() => ProposalActionsModule),
     forwardRef(() => VotesModule),
+    ProposalActionRolesModule,
     GroupsModule,
     ImagesModule,
     RolesModule,
